requirements.txt
appnope==0.1.2
asgiref==3.3.1
astroid==2.5.1
backcall==0.2.0
decorator==4.4.2
Django==3.1.6
djangorestframework==3.12.2
ipython==7.21.0
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.18.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
psycopg2==2.8.6
ptyprocess==0.7.0
Pygments==2.8.0
pylint==2.7.2
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
typed-ast==1.4.2
wcwidth==0.2.5
wrapt==1.12.1
